Added build scripts and ChezScheme submodule.
This commit adds two scripts explained in the README file and - most importantly -
a git submodule to automatically download the ChezScheme repo for building directly
ChezScheme from source which is needed for chez-exe.

It is also possible to build chez-exe with an already installed ChezScheme compiler:
note that the process has been tested on Void Linux (x86_64) but it SHOULD work
on other distros as well.